Move binding methods to own file
